feat(ui5-segmented-button-item, ui5-toggle-button): migrated to TypeScript
